feat: add route to delete book and reactoring.

+++ src/domain/application/Book/DeleteBook/DeleteBook.ts
@@ -4,6 +4,7 @@ import UseCase from "@/domain/application/UseCase";
 import Repository from "@/domain/core/Repository";
 import DeleteBookOutputBoundary from "./DeleteBookOutputBoundary";
 import { DBOutputBookData } from "@/domain/application/@types/BookTypes";
+import EntityNotFoundError from "../../Errors/EntityNotFoundError";
 
 export default class DeleteBook implements UseCase<{ id: string }, boolean> {
   constructor(readonly repository: Repository) {}
@@ -11,12 +12,15 @@ export default class DeleteBook implements UseCase<{ id: string }, boolean> {
   async execute(
     inputData: InputBoundary<{ id: string }>,
   ): Promise<OutputBoundary<boolean>[]> {
-    const { id } = inputData.get();
+    const { id: _id } = inputData.get();
 
-    await this.repository.delete({ id });
+    const deletedBook = await this.repository.delete({ _id });
+    if (!deletedBook) {
+      throw new EntityNotFoundError("Book");
+    }
 
     const dbBook: DBOutputBookData | null = await this.repository.getOne({
-      id,
+      _id,
     });
 
     return [new DeleteBookOutputBoundary(dbBook)];
